SE AGREGO LA FUNCIONALIDAD PARA RESTRINGIR LOS MODULOS DE ACUERDO AL PERMISO

+++ resources/views/livewire/permission/permission-livewire.blade.php
@@ -4,7 +4,9 @@
                 <div class="card-header">
                     <div class="row d-flex justify-content-between">
                         <div class="col-md-1">
+                            @can('agregar-permiso')
                             <button class="btn btn-success btn-md" wire:click="$toggle('openModalNew')"><i class="fas fa-plus-square"></i></button>
+                            @endcan
                         </div>
                         <div class="col-md-4">
                             <div class="input-group mb-3">
@@ -25,7 +27,9 @@
                                 <th>#</th>
                                 <th>NOMBRE</th>
                                 <th>DESCRIPCIÓN</th>
+                                @can('funciones-permiso')
                                 <th>ACCIÓN</th>
+                                @endcan
                             </tr>
                         </thead>
                         <tbody>
@@ -34,16 +38,20 @@
                                     <td class="text-center">{{ $loop->iteration }}</td>
                                     <td>{{$permission->name}}</td>
                                     <td>{{$permission->description}}</td>
-                                    <td class="text-center">
-                                        {{-- @can('register') --}}
-                                        <button class="btn btn-warning btn-sm" wire:click="edit({{ $permission->id }})"><i class="fas fa-pencil-alt"></i></button>
-                                        {{-- @endcan --}}
-                                        @if ($permission->estado == 1)
-                                            <button class="btn btn-danger btn-sm" wire:click="$dispatch('confirmDelete', {{ $permission->id }})"><i class="fas fa-trash"></i></button>
-                                        @else
-                                            <button class="btn btn-primary btn-sm" wire:click="$dispatch('confirmDelete', {{ $permission->id }})"><i class="fas fa-history"></i></button>
-                                        @endif
-                                    </td>
+                                    @can('funciones-permiso')
+                                        <td class="text-center">
+                                            @can('editar-permiso')
+                                            <button class="btn btn-warning btn-sm" wire:click="edit({{ $permission->id }})"><i class="fas fa-pencil-alt"></i></button>
+                                            @endcan
+                                            @can('eliminar-permiso')
+                                            @if ($permission->estado == 1)
+                                                <button class="btn btn-danger btn-sm" wire:click="$dispatch('confirmDelete', {{ $permission->id }})"><i class="fas fa-trash"></i></button>
+                                            @else
+                                                <button class="btn btn-primary btn-sm" wire:click="$dispatch('confirmDelete', {{ $permission->id }})"><i class="fas fa-history"></i></button>
+                                            @endif
+                                            @endcan
+                                        </td>
+                                    @endcan
                                 </tr>
                             @endforeach
                         </tbody>
